+++ inference/inference_requirements.txt
@@ -1,1 +1,2 @@
 pandas>=2.1.4
+scikit-learn>=1.2.2
