+++ k8s/service.yaml
@@ -10,4 +10,4 @@ spec:
   ports:
     - protocol: TCP
       port: 80
-      targetPort: 5000
+      targetPort: 80
