--- web-app/src/App.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import PredictionApp from './PredictionApp';
-
-function App() {
-  return (
-    <div className="App">
-      <PredictionApp />
-    </div>
-  );
-}
-
-export default App;
