--- web-app/src/index.css
@@ -1,38 +1,0 @@
-body {
-  background-color: #f8f9fa;
-}
-
-.card {
-  max-width: 500px;
-  margin: auto;
-}
-
-h1 {
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-  color: #343a40;
-}
-
-h3 {
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-}
-
-.btn {
-  font-size: 1.1em;
-}
-
-h3.text-info {
-  font-size: 2em;
-  margin-top: 20px;
-}
-
-.positive {
-  color: green;
-}
-
-.neutral {
-  color: blue;
-}
-
-.negative {
-  color: red;
-}
