+++ web-app/web_requirements.txt
@@ -1,2 +1,3 @@
 Flask>=2.1.1
 flask-cors>=3.0.10
+pandas>=2.1.4
